% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{getExternalPtr}
\alias{getExternalPtr}
\alias{getDuplicateMethod}
\alias{getCoerceMethod}
\alias{setDuplicateMethod}
\alias{setCoerceMethod}
\title{Get/Set the properties of the wrapper of a vector pointer}
\usage{
getExternalPtr(x)

getDuplicateMethod(x)

getCoerceMethod(x)

setDuplicateMethod(x, value = c("duplicate", "sameObject", "error"))

setCoerceMethod(x, value = c("coerce", "error"))
}
\arguments{
\item{x}{an ALTREP vector, the return value of `wrapPointer` function}

\item{value}{character, the value you want to set.}
}
\value{
`getExternalPtr`: the externalPtr object associated with the wrapper
`getDuplicateMethod` : character, the duplicate method
`getCoerceMethod` : character, the coerce method
}
\description{
Get/Set the properties of the wrapper of a vector pointer
}
